requirements.txt
openai-whisper==20240930
setuptools-rust==1.10.2
pytest==8.3.5
httpx==0.28.1
python-multipart==0.0.20
fastapi
uvicorn
sqlalchemy